solved all questions, and the visualizer is changed to be a notebook
